+++ src/deltachat_cursed/account.py
@@ -1,3 +1,5 @@
+from typing import List
+
 from deltachat import Account as _Account
 from deltachat.capi import ffi, lib
 
@@ -10,3 +12,13 @@ class Account(_Account):
         return lib.dc_array_get_cnt(
             ffi.gc(lib.dc_get_fresh_msgs(self._dc_context), lib.dc_array_unref)
         )
+
+    def get_messages(self, chat_id: int) -> List[int]:
+        """Return list of messages ids in the chat with the given id."""
+        dc_array = ffi.gc(
+            lib.dc_get_chat_msgs(self._dc_context, chat_id, 0, 0), lib.dc_array_unref
+        )
+        return [
+            lib.dc_array_get_id(dc_array, i)
+            for i in range(0, lib.dc_array_get_cnt(dc_array))
+        ]