+++ src/deltachat_cursed/event.py
@@ -17,21 +17,25 @@ class AccountPlugin:
     def __init__(self, account: Account) -> None:
         self.account = account
         self.chatlist_monitors: Set[ChatListMonitor] = set()
-        for chat in self.account.get_chats():
+        for chat in self.get_chats():
             self.current_chat = chat
             break
         else:
             self.current_chat = None
 
+    def get_chats(self) -> List[Chat]:
+        return [chat for chat in self.account.get_chats() if chat.id >= 10]
+
     def add_chatlist_monitor(self, monitor: ChatListMonitor) -> None:
         self.chatlist_monitors.add(monitor)
-        monitor.chatlist_changed(self.get_current_index(), self.account.get_chats())
+        chats = self.get_chats()
+        monitor.chatlist_changed(self.get_current_index(chats), chats)
 
     def remove_monitor(self, monitor: ChatListMonitor) -> None:
         self.chatlist_monitors.discard(monitor)
 
     def chatlist_changed(self) -> None:
-        chats = self.account.get_chats()
+        chats = self.get_chats()
         if self.current_chat not in chats:
             if chats:
                 self.current_chat = chats[0]
@@ -46,13 +50,13 @@ class AccountPlugin:
             m.chatlist_changed(index, chats)
 
     def select_chat(self, index: Optional[int]) -> None:
-        chats = self.account.get_chats()
+        chats = self.get_chats()
         for m in self.chatlist_monitors:
             m.chat_selected(index, chats)
         self.current_chat = None if index is None else chats[index]
 
     def select_chat_by_id(self, chat_id: int) -> None:
-        chats = self.account.get_chats()
+        chats = self.get_chats()
         chat = self.account.get_chat_by_id(chat_id)
         index = chats.index(chat)
         for m in self.chatlist_monitors:
@@ -60,7 +64,7 @@ class AccountPlugin:
         self.current_chat = chat
 
     def select_next_chat(self) -> None:
-        chats = self.account.get_chats()
+        chats = self.get_chats()
         if self.current_chat is None:
             if chats:
                 self.select_chat(len(chats) - 1)
@@ -72,7 +76,7 @@ class AccountPlugin:
             self.select_chat(i)
 
     def select_previous_chat(self) -> None:
-        chats = self.account.get_chats()
+        chats = self.get_chats()
         if self.current_chat is None:
             if chats:
                 self.select_chat(0)
@@ -83,10 +87,9 @@ class AccountPlugin:
                 i = 0
             self.select_chat(i)
 
-    def get_current_index(self) -> Optional[int]:
+    def get_current_index(self, chats: List[Chat]) -> Optional[int]:
         if self.current_chat is None:
             return None
-        chats = self.account.get_chats()
         return chats.index(self.current_chat)
 
     @account_hookimpl